add confirm modal for deletion

+++ frontend/src/components/Modals/MoreModal/index.tsx
@@ -1,5 +1,5 @@
 import React, { useRef } from "react";
-import { useAppDispatch, useAppSelector } from "../../../hooks/redux";
+import { useAppDispatch } from "../../../hooks/redux";
 import { useOutsideModal } from "../../../hooks/useOutsideClose";
 import {
   deleteBoard,
@@ -11,6 +11,9 @@ import {
   onModalEdit,
   onToggleModal,
 } from "../../../store/reducers/modalReducer";
+import { confirm } from "react-confirm-box";
+import {Options} from 'react-confirm-box/dist/types'
+import ConfirmModal from "../ConfirmModal";
 import "./MoreModal.scss";
 interface MoreModalType {
   id: string;
@@ -22,12 +25,15 @@ const MoreModal = ({ id, type, setIsOpen }: MoreModalType) => {
   const dispatch = useAppDispatch();
   const modalRef = useRef<HTMLDivElement | null>(null);
   useOutsideModal(modalRef, setIsOpen);
-  const onDeleteHandle = () => {
-    if (type === "column") {
-      dispatch(deleteColumn(id));
-    } else if (type === "board") {
-      dispatch(deleteBoard(id));
-      dispatch(setSelectedBoard(null));
+  const onDeleteHandle = async (options: Options) => {
+    const result = await confirm("Are you suree ?", options);
+    if (result) {
+      if (type === "column") {
+        dispatch(deleteColumn(id));
+      } else if (type === "board") {
+        dispatch(deleteBoard(id));
+        dispatch(setSelectedBoard(null));
+      }
     }
 
     setIsOpen(false);
@@ -45,7 +51,7 @@ const MoreModal = ({ id, type, setIsOpen }: MoreModalType) => {
   return (
     <div ref={modalRef} className="MoreModal">
       <button onClick={onEditHandle}>Edit</button>
-      <button onClick={onDeleteHandle}>Delete</button>
+      <button onClick={() => onDeleteHandle(ConfirmModal)}>Delete</button>
     </div>
   );
 };
